Related to issue 27887.Boolean filters are advanced criterias
When filtering by a boolean in a grid the criteria received by the datasource
is an AdvancedCriteria. New isInScope method created for boolean fields based
on the version for BigDecimals to cope with this type of criterias.

--HG--
extra : transplant_source : i%0A%F8%10s%3D%DB%ED%84%CBE%D6%0C%98%C7%B9%28%804%AC