Fixes issue 27957: Grid button don't get stuck after autosaving new record
In the fix [1] of this issue [2], a change was made to force reselecting a new record even if it was already selected, because otherwise when a record was created in grid mode while a filter was applied, when the record was saved the selection was lost.

The problem is that that change should only be applied when the editCompletionEvent is 'programmatic'. In the other cases (i.e. 'tab', 'enter', etc), the focus goes to another record, so it did not apply to [2]. In fact when the editCompletion event is not 'programmatic' not only reselecting the record was not needed, but also it caused the grid buttons to get stuck.

[1] https://code.openbravo.com/erp/devel/pi/rev/dfda65fa925f89a21eee64f35d7aadb252c90dab
[2] https://issues.openbravo.com/view.php?id=26702