Fixed issue 27974.Do not filter queries by 31-12-9999
Some queries have been modified so when the costing rule is not fixing
backdated transactions it is not needed to filter by 31-12-9999.

This filter was causing a SQL Exception that caught by the Costing Background
but no rollback was done. This is also fixed.

--HG--
extra : rebase_source : 56f8ff65b0ac5211896051c00453dcec44d978d0