Fixed issue 27974.Do not filter queries by 31-12-9999
Some queries have been modified so when the costing rule is not fixing
backdated transactions it is not needed to filter by 31-12-9999.

This filter was causing a SQL Exception that caught by the Costing Background
but no rollback was done. This is also fixed.

--HG--
extra : transplant_source : %96%2B%FASn%18%A7%CC%17%A1%B9%D9%40%A4iG%B5%CA8%0B