Fixed bug 27944: Not Stocked BOM lines are printed without price in orders.
Print sales order report was modified to take into account price and total amount for order lines reffering products with some bill of material defined. Now it is possible to delete invoice header.
This change affects also the purchase flow so now when printing an order, price and total of the exploded BOM lines are included, and the total match the total gross amount of the order.