Related to issue 28029: Prevent extra negative stock adjustment line
AverageCostAdjustment it was creating an extra negative stock adjustment line when there was already an adjustment line of negative stock. This was happening on a receipt with previuos stock in negative, that leaves the stock still in negative after the completion of the receipt.
To prevent this and extra check has been added in AverageCostAdjustment: If baseCAL is of type negative stock, then dont add a new negative stock line.

--HG--
extra : transplant_source : %C68%9C%EB/%11w%F9%C1%A9%27%B9%D2%A6F%C8%B6%3C%A2%EC