Related to issue 28029: Prevent extra negative stock adjustment line
AverageCostAdjustment it was creating an extra negative stock adjustment line when there was already an adjustment line of negative stock. This was happening on a receipt with previuos stock in negative, that leaves the stock still in negative after the completion of the receipt.
To prevent this and extra check has been added in AverageCostAdjustment: If baseCAL is of type negative stock, then dont add a new negative stock line.

--HG--
extra : rebase_source : 78bc22bd69adde55e167e5d0d8daf9366774cd31