Fixes issue 28099: '.' works as date separator
If a dot was used as date separator, the dates were shown empty in the UI. This was caused by this implementation of the isSeparator function:

  isSeparator: function (str, position) {
    return str.charAt(position) === '-' || str.charAt(position) === '\\' || str.charAt(position) === '/';
  },

The list of possible separators was hardcoded, and it did not take into the account the dot. This has been fixed by also checking the value returned by the OB.Utilities.Date.getDateSeparator function, that given a date format it returns its separator.

--HG--
extra : rebase_source : 868eab6f1a0a4fb4b14049c9fadf1b30b307f0a8