Related with issue 28119: DataSourceServiceProvider.getDataSource refactor #1
The DataSourceServiceProvider.getDataSource method is going to be refactored before fixing 28119, because its logic is very messy.

In this first part of the refactor the obtention of the DataSource and the obtention of the DataSourceService is going the be separated. The first part of the method now is in charge os retrieving the DataSource, and the second part obtains the DataSourceService associated with the DataSource.

--HG--
extra : rebase_source : 8f6003e6e383df25a1308f6486eb50da3dcd04a0