Fixes Bug 27998: Negative Stock Correction adjustments now are properly created
Due to Issue 27946, each time that an adjustment is created it is checked if a Negative Stock Correction adjustment is needed, and if so, it is created. So now, in Costing Server, before creating any Negative Stock Correction it is checked if another Cost Adjustment has been created, and if it is, this step is skipped.

--HG--
extra : transplant_source : N%ECqz%F2%40y%D46XF%B5%A8%BFp%85%C1%C3%DD%A2