[hqlpickandexecute] Send proper tabId in request
Before that last pi merge, the tabId being sent in the request was the tabId of the button owner tab. After the merge, it was the tabId of the pick and execute tab. Now it has been fixed so the tabId sent it the tabId of the button owner tab.