Fixes bug 28113: Use costing Price to calculate Adjustment Amount
In some cases the transaction price it was not being properly calculated in order to calculate the negative stock correction amounts. Starting from now the Net Unit Price field of Costing tab will be used in Cost Adjustments. To get the transaction price this Net Unit Price will be used.