Fixes issue 25811: No unneeded requests are done when the grid is refreshed
This changesets handles the case when the grid is refreshed after having been scrolled down, without no records selected. When this happens, the refreshingWithScrolledGrid flag is set to true. This will be
 used in the fetchRemoteData function to help prevent a grid redrawal that would result in an unneeded datasource request.

Also, all the flags used in this development are cleared in the callback of the filterData function done in OBViewGrid.refreshGrid.