Related with issue 25811: Adds responseAction that allows to refresh the grid
After the fix of [1], the whole grid is no longer being refreshed after executing a standard process, but only the selected record. If the process adds or deletes records from the tab owner of the button, then it include the refreshGrid function in the list of responseAction of the standard process.

This refreshGrid function does not have any parameters, and it must be added to the responseActions like this:

  responseActions.put("refreshGrid", new JSONObject());

[1] https://issues.openbravo.com/view.php?id=25811