Fixed bug 27999: Product's image created on create variants can not be changed.
Actual creating variants process makes a copy from the entire product's object to create the new variants and the new variant's images are the original product's image so it is not possible to update them. This behaviour was changed to also make a copy from original product image, save it and then associate it with the new variant. Now it is possible to update this new image's field as it is not related with any other variant or product.
Also a module script was created to fix this problem in already created variants through the mentioned process. When executing an ant update database, images for products with generic products associated are reinserted and then are associated with them.