Fixed issue 28098: Loading text is shown vertically when using japanese language.
When the languaje was japanese, the text showing "loading..." in japanese charcteres was shown vertically.
I have changed the width of the layout from 1 to 100 and added an align to the right.
The problem was that each japanese charcter was treated as an independent word, and when there was more than one word,
automatically it was posted down the other word. So, we saw it vertically.