Related with bug 25811: Dont refresh the subtabs always after refreshing form
Not always when the OBViewForm.refresh function should its children tabs be refreshed. Sometimes they should be refreshed (i.e. when pressing the refresh toolbar button of after a process is executed), but other times they should not (i.e. when the form is refreshed because a record has been saved in one of its sutabs).