Related with bug 25811: Set flags to prevent multiple requests in proper placei
The refreshingWithRecordSelected and refreshingWithScrolledGrid, which are being used to prevent unneeded multiple requests, were being set to true in the wrong place. They were being set to true before ac
tually starting to refresh the grid, and that result that under some circumstances not even the first datasource request was done.

Now they are set to true right after starting the refresh, just after calling OBViewGrid.filterData. They are being reset in the same place as before, in the callback of the filter data function.