Related with bug 25811: Process view should be refreshed after invoking process
After a manual process is invoked, the closeProcessPopup function is called. This function is in charge, among other things, of refreshing the tabs. The problem was that it was doing it taking as the primery view the view that was active when the process was invoked, not the view that defines the button that opens the process, which is the one that most probably be affected by the process.

Fixing this also prevents an unneeded request when invoking a process from the subtab of the tab that defines the process (i.e. booking an order from the sales order line tab).