Fixes bug 28220:FK no longer filtered without applying selected parent criteria
The problem was the following. If a tab was configured to use lazy filtering, when one of its foreign keys is filtered the criteria is obtained from the current values of the filter editor, instead of directly calling the getCriteria function of the grid.

The problem is that the criteria was not being passed to the OBViewGrid.convertCriteria function, which among other things adds the criteria to filter by the record selected in the parent tab. The result was that the dropdown contained way more records than it should, and that the query to retrieve the records could have very bad performance if the table associated with the tab contained many records.

--HG--
extra : transplant_source : x%DES%E9%FCrS%11Wt%A8%18%F3%CC%E6%20%F2X%9DA