Fixes issue 28214: _selectedProperties parameter work in JSON REST webservice
In a call to the Openbravo JSON REST webservice, the _selectedProperties parameter allows the user to specify the set of properties whose values should be returned by the webservice. This property was working properly when an ID is specified in the request, but not if it is not specified.

This stopped working because of this issue [1]. [1] implements a new method to handlle the case when the ID is not specified to improve the performance, but its missing the handling of _selectedProperties.

[1] https://code.openbravo.com/erp/devel/pi/rev/c171ecac2673