Fixes issue 28212: Prevent NPE in FormatUtilities.replaceJS
In this changeset [1] the FormatUtilities.replaceJS was updated to make it secure against cross site scripting. The problem was that before the change the method returned null if the strIni parameter is null, but after the change a NullPointerException is thrown. Added a check to prevent this.

[1] https://code.openbravo.com/erp/devel/pi/rev/244f648e594e

--HG--
extra : rebase_source : 2187856beecfd41239a5e1f884d2213348f2d015