Fixes issue 28229: Lazy Filtering works properly in P&E windows
There was a problem when the button to apply the current filters (only shown when lazy filtering is configured for a tab) was pressed in a P&E window. If the filters were not changed before pressing the button, the grid.refreshGrid function was invoked, and this function was not defined for P&E windows. This function has been defined now. Also, a flag has been added to detect when the filterClause is removed, to allow refreshing the grid in that case even if the filter editor is not changed.