Fixes issue 28018:Null pointer exception when using obuisel_selectors
An uncaught null pointer exception when using obuisel_selectors based on datasource.
The exception is thrown on the SelectorDefaultFilterActionHandler class.
In some cases this happens when this kind of selector definition does not use a table but a datasource.


This has been fixed and now it can obtain the name of a table based on datasource too.