Related with bug 25811: Dont show warn message if _targetRecordId param is null
In the changeset c5de0b748930 a wanrning message was added when a datasource request with the _targetRecordId but without the _directNavigation parameter was detected. The check was not properly done, and was not discarding the cases when the _targetRecordId was present in the parameter list but its value was null.

--HG--
extra : rebase_source : 6b734f82faf7d3a91f0b69fec6992cf7d302b7d8