Fixes issue 28178: Local edits are discarded when the grid is refreshed
If there is a validation error when editing the row of a grid, that change is stored in the local edits of the grid, but not persisted in the database. The problem was that when the grid was refreshed the updated records returned by the datasource where being overwritten with the grid local changes.

This behaviour was wrong, when the grid is refreshed all changes should be discarded. To fix this, the local edits of the grid are discarded when the grid is refreshed, right after invalidating the cache.

--HG--
extra : rebase_source : f0bf1bb346287f64aee0391ad84d11794c2dee66