Fixed issue 28169: readOnly and readOnly logic fails in images.
In image fields it did not work readOnly property nor readOnly logic. When readOnly was chosen in the image field, it was possible
to delete or upload it. When readOnly logic was active it was possible to do both things too.
The problem was in the ob-formitem-image.js.
For fixing the first problem I have changed the word "readOnly" for "disabled" in a condition, because the logic was not taken well before.
Now, when the checkbox readOnly is chosen, it is not possible to delete nor upload an image, as expected.
For fixing the second problem I have overwritten the "setDisbaled" function in this subclass. Now, when an obect is set as disabled, the delete and upload
buttons will be hidden, so, when an image is set as readOnly whit a readOnly logic condition, now those buttons are not visible.

--HG--
extra : rebase_source : acaab22fa0250c430579de9221671b2e6c58a220