Fixes issue 28178: Disable refresh button while the view is being refreshed
Before the fix, the user could press the refresh toolbar button several times before the response of the first one was received. This could result in the server being overloaded. To prevent this, now when
then user clicks on the refresh toolbar button, the button is disabled until the response is received.

To ensure that the refresh button is always enabled even if an error happens, the button is being enabled in several places:
- If the refresh process is successful: In the callback of the refresh function.
- If there is an error between the refresh function is called and the request is actually sent: In a try/catch block
- If the server returns with an error code: In the RPCManager.handleError function

--HG--
extra : rebase_source : e177ced6230d494a5b69fce4fcbdf5ce2727706e