Related with issue 28188: Handle better the records with validation errors
When a record is saved and it has a validation errors, its _hasValidationErrors flag is set to true. At no point this flag was set to false when the validation errors were solved (either by cancelling the
changes or by saving the record again after fixing the data).

Now the handling of the validation errores is done using the addRecordToValidationErrorList and removeRecordFromValidationErrorList functions. This functions not only set to true or false the _hasValidatio
nErrors flag, but also keep a list of the ids of the records that have validation errors. This allows to know at any point what records have validation errors without having to iterate the whole grid local
 data. Also a function has been added to check if the grid has any validation errors.