Fixes issue 28188: Backs out original fix, another approach is used
When the grid has data that has not been saved due to having validation errors, instead of allowing the user to refresh the record and get rid of the local changes, now the refresh button is disabled. The cancel button can be used to restore the data to the state it had before the validation errors, and once there are no more invalid records the refresh button will be enabled again.