Fixes Issues 27820:Error when posting Payment/Transaction/Reconciliation referencing a cash vat invoice linked to several orders
Issue is a regression due to related fix for issue 26571. Cast Vat lines
were posted as many times the no. of orders linked with cash vat invoices.
So now posting is based on the payment details so the relevant cash vat line
associated with it is posted though it is looped as many times as the no. of
orders linked.

--HG--
extra : transplant_source : %81%E11%E2%AF%2Cy%E0%BF%3C%FE.-Yd%ADT%B4%DB-