Fixed issue 28242: Load grid even with inconsistent default view info
Before the previous changeset was applied, it was possible to have a preference that specified that a certain grid had a default saved view, but that the referenced default saved view did not actually exists. When this happened the grid was not loaded until the user pressed the Refresh button.

To fix this, now the inconsistent default view info is detected and the grid is loaded once it is detected that the default view does not actually exists.

--HG--
extra : rebase_source : 2f413332905f3f0513cacc684edc11bbc25cb226