Fixes issue 28276: AccessLevelTest does not take into account HQL based tables
The AccessLevelTest.testADataAccessLevel tests if the current client/org of all tables in the database is valid for the access level defined for that entity. It does so my executing a query to the database table associated with the ad_table.

ad_tables based on manual datasources or on hql queries do not have database tables associated, so this test should not consider them. This was already working properly for ad_tables based on manual datasources, but not for datasources based on HQL tables.