Fixes issue 28246: Grid parameters are loaded even if they have display logic
The problem was that the grid parameters were only being automatically loaded when they did not have display logic. If they had display logic the user had either to programatically load the grid, or change the grid filter editor.

To fix this, now it is detected when a grid is shown for the very first time (in that case its data property will not be a ResultSet, as it has not been loaded yet), and it is loaded then.

--HG--
extra : rebase_source : cc8ed6066ce68d1b056b6662b3ffb7f2f5456ed6