Fixes issue 28231: Form is not scrolled if it does not have any editable fields
The problem was that when switching from grid to form view, if the form view did not have any editable fields the focus would be placed in the first Audit field, and the scroll would be moved so that the Audit field is shown in the top of the screen. This would hide all the read only fields.

To fix this, when the switchFormGridVisibility is invoked it is detected if the focused item is an Audit field, and in that case the scroll is moved to the top so that the first batch of form fields is visible.