Related with bug 28246: Filter expression not reset after clearing the filters
The following problem has been detected while doing the QA for issue 28246: If a grid parameter has a display logic and a filter expression, then when the user clears the filter the filter expression will be momentarily removed, and then reset again.

This happened because there was some code that tried to guarantee that the grid was loaded the first time that is was made visible and that had the filter expression applied, but was not checking properly that the grid was not loaded yet. This resulted in reapplying the filter expression each time the grid is redrawn.

This has been fixed by checking properly if the grid has already loaded its data (in that case its data property will be an instance of ResultSet).

--HG--
extra : rebase_source : dbe90227c8c4557f4935bc31b048f6e002d789f1