Fixes issue 27829: Column order definition in HQL tables does not matter
There was already a mechanism in place to support that the order of the columns in the HQL query is different from the order of the columns in the application dictionary. The problem was that this mechanism was not working properly when the column aliases had uppercase characters, due to a missimg .toLowercase statement in the Entity.getPropertyByColumnName method. Due to this bug the HQLDataSourceService was not able to find the properties of the entity based on the column alias, and was instead using the index of the column to  pick the value from the result array, making the column order relevant.

--HG--
extra : rebase_source : 2d09f325c1148246a0704ba83b6b43256ae4cae3