Fixed issue 27787: Create new icon is shown when windows are readOnly.
When the window/tab access for roles was readOnly, the create new icon was shown by the name of the windows, in the recent views layout.

For fixing this the "MenuManager.java" has been changed.
The logic for the window/tab access for the roles has been created. This is for knowing if the window/tab access for roles is read only or not.
Now, if a role has read only access the "create new" icon is not shown in the recent views layout.
It is good to know that the role access data is cached in the quick launch menu and in the recent views layout, so at first the "create new"
green icon by the name of the window in the recent view layout still appears. It is not a problem because when the window is cached again the
correct status is shown.