Fixes issue 28187: Row is saved even with ProductCharacteristics column hidden
Product characteristics are somewhat special in the sense that the value returned by the FIC in columnValues is different from the value that is finally stored in the database. If a product has a characteristic 'X' with a value 'Y', the FIC returns this json object:
{ characteristics: { X:'Y'},
  dbValue: 'X : Y'
}

In the database the stored value should be the dbValue of this json object. This replacement is done in th setValue function of the product characteristics formitem. The problem is that this function is not invoked, and cannot be invoked, if this column is hidden in the grid. When the function is invoked, instead of storing string contained in the dbValue property in the database, the whole json object returned by the FIC is sent to the datasource and a validation exception is thrown. To fix this, the value is going to be fixed in the JsonToDataConverter.convertJsonToPropertyValue method. To be able to discern that the property belong to a Product Characteristics column, the ProductCharacteristicsDomainType class has been created and has been associated with the Product Characteristics reference. This class has a method called fixValue, that detects the case when the provided value is not correct, and fix it.

--HG--
extra : rebase_source : b12105219a9fd27853a8fae0b68a12dfe9522660