Fixed issue 27436: A broken trigger needs to be removed.
There was a trigger called AD_TREENODE_MOD_TRG that was not working well. It was created to control if a module was in development or not
and if it was not in development, it couldn't be  possible to move it arround the nodes on the tree view in the "Menu" tree.
This feature is handled by the MenuTreeOperationManager.class, and it works well.

For fixing the issue the AD_TREENODE_MOG_TRG trigger has been removed from the database.