Fixed issue 28281: When using lazy filtering, funnel icon is not hidden.
When the lazy filtering is enabled, the funnel icon for clearing the filters was not hidden after clicking it.
The problem was in the "ob-grid.js" file.
The "this.filterImage.hide()" line has been added to the "clear filter" function for hidding the funnel icon. Now, when the lazy filtering is enabled
and the funnel icon is clicked for clearing the filters, the funnel icon disappears as expected.