Fixed issue 27877: No rollback is called in connection error catch.
It seemed to be possible to have a connection leak caused in the "UsageAudit.java" class.
That was because at the final function of this class in a try catch block there was no rollback done.
After some tests it seems not to be any connection leak even without the rollback. But the proper code
to make the rollback has been created, because it seems to be better that way.
For developing this, a new method has been created "releaseConnection". This method is called in the three
try catch blocks of the method "auditActionNoDal".
Now, if one of those excepcions are reached, a rollback will be done.