Fixes issue 28231: Form is not scrolled if it does not have any editable fields
There was a problem with the previous fix: it worked when a record was opened from the grid view to the form view, but then it did not work if the user moved to the previous or next record while in form view.

To fix this, the logic that handled the original problem has been moved from OBStandardView.switchFormGridVisibility function to the OBViewForm.setFocusInForm function. The flow goes through the latter function each time a new record is loaded. In particular the form has to be scrolled down only when the record in being initializaed in the form, that's why a flag has been added to ensure that the code is only executed when the setFocusInForm function is executed from within the retrieveInitialValues function.