Fixed bug 27809: The Wrong sales invoice alert can not be properly deactivated.
Method existsAlertRule on Alerts For Wrong Invoices module script xsql file was changed to not take into account if wrong invoice alert is active or not so when module script is executed and the mentioned alert is deactivated it is recognized as an existent rule and it is not inserted again as active.
Now it is possible to deactivate The Wrong sales invoice alert and conserves its status, no matter an update database is executed.