Related with issue 27787: Create new icon is shown when windows are EditOnly
The case of the tab's UI Pattern put as "Edit Only" was not taken in the first fix of the issue.
So that case is fixed now.
When the tab's UI pattern was "Edit Only", the small green icon for "create new" was shown in the
recent views layaout.
For fixing this the "ob-utilities.js" has been changed.
Inside the "{OB.Utilities.openView}" function there was a condition for setting the "singleRecord" value
 (that is the Edit Only attribute) of the tab
to true or false, wich was dependent on other condition. It was not correct, because the "singleRecord" value does not depend
on the other condition for being true nor false.
The condition that evaluates the value of "singleRecord" has been moved outside of the other condition.