Fixes Issue 28234: Check in Costing Rule Validation that the org has Currency.
If the organization has no currency defined, the costing process will use the currency defined in the client. This could be wrong if the currency for the transactions of this organization should be different that the currency of the client, so before validating a costing rule the organization of it should have a currency defined.

--HG--
extra : transplant_source : %A3c%C2%9A%EE%E3%F2%7D%0E%838%F4%EB%A2%FC%7F%8E%97%D1%2C