Fixes bug 28230 Null Pointer Exception is raised when trying to post a goods receipt
When a goods receipt with a product of type Service, which has a cost of type Standard, is posted, the DocInOut class calls the getStandardCost method from CostingUtils class with the currency of the current organization. If the organization has no currency, the client's currency will be used. Otherwise, a null pointer exception will be raised.