Fixed bug 28156: Completing invoice with 0 grandtotal, finalsettlement is null.
Extension point executed when completing invoices stored procedure is finished, was modified to not only update isPaid field as Y, but also to fill the finalsettlement with the current date which is when the document with grandtotal as 0, is being completed.
Now when completing an invoice with grandtotal as 0, also the finalsettlement field will be filled, avoiding the invoice is processed by the fin_paymentmonitorprocess unnecessarily. This change affects purchase and sales flows.

--HG--
extra : rebase_source : 22c2b7acefa510b55a7697d094fc5b98f6522860