Fixed issue 26386: Process buttons are not shown in the tree grid view.
Process buttons are not shown in the tree grid view if no records are selected in
the standard grid view.

It has been solved by controlling the current view: tree grid view or grid view.
Now, Process buttons are shown in the tree grid view if no records are selected in
the standard grid view.