Fixes Issue 28157: Landed Cost takes into account exchange rate at invoice level
If an exchange rate different from the System is added in Purchase Invoice and this Invoice is matched in a Landed Cost, a new record will be created in 'Matched Amount' tab under 'Landed Cost Cost' tab of 'Landed Cost' window. This new record will have the new flag of 'Is Conversion Matching' as true, and the amount will be the difference obtained from using the Invoice exchange rate compared to the one of the System.
This will create a new Cost Adjustment of source 'Landed Cost' with the same amount.

--HG--
extra : transplant_source : %28%FF%19%FE4%5E%10%2B%AD%99%07%7CG%3DC%0B%9EtW%14