Fixed issue 28047:An exception is shown in console.
There was an exception that was ignored (this is a correct behaviour), but it was shown in the log.
This exception appeared when, in cube definition window a cube Report Dimension was created. Then the log showed
the exception, but the Cube Report Dimension was correctly created. So, the exception needed not to be shown.

For fixing this the "AdvancedQueryBuilder" java class has been modified.
The exception was thrown before because the method <getPropertyByColumnName(String ColumName)> was called
and this method always checks the "isNotNull". When the variable was null, the message and the exception was catched
and shown.
For avoiding this, the method <getPropertyByColumnName(param, checkIsNoNull)> has been called instead the other one.
This method, has an input parameter called "checkIsNotNull", putting this to false, avoids the evaluation of the function
"isNotNull", so, if the variable is null it doesn't matter and the message and the exception shown before does not appear now.
The problem is that we get a  "nullPointerException" if we do nothing else more.
For avoiding this, a condition has been added to the code in the "AdvancedQueryBuilder" java class. If the prop variable is
null, the paramValue variable gets the empty string value. And as the value for paramValue variable cannot be reached from
the request, it is get from the context.

--HG--
extra : rebase_source : e4c9b5fd79a3c045fa1e33d41700ac95f28d1dfc