Fixed bug 27280: Commission payment details are duplicated in some cases.
Generate commision process was changed to exclude the same records when retrieving data to be inserted in the payment detail table. This information comes duplicated due to the payment is divided in two records as the invoice has fractionated payment terms.
Thus the stored procedure was changed only in those cases when the commission basis document is an invoice, and the information related with the payments and its products are joined in the same query, avoiding duplicated tuples.
Now there are not duplicated fields when processing commissions for invoices with fractionated payment terms. Also a module script was developed to correct inconsistent data. When executing an ant update database duplicated data in commission detail table will be deleted, also amounts and quantities will be updated in commission amount and commission payment tables with the correct information.