Fixed issue 28384: Console error when logging out from WebPOS if there are orders pending to be paid
The code executed in the postCloseSession function has been refactored and simplified.
Before, a find was executed to retrieve pending orders. For each order, two removes were executed: the first remove was wrong and was causing a console error. The second remove was removing the order from the database.
After the fix, the find has been replaced by a removeAll with the same criteria, and all pending orders are deleted in a single query.

--HG--
extra : rebase_source : a78abe610823193808e30640bb056e3ae0d83d7b