Fixed issue 28334: The default value on UOM is not taken.
When the user level is chosen as "organization". The default value set in the Unit of Measure is not
taken when the user creates a new product. Instead of the default value of UOM the first one appears.

For fixing this the "LoginUtils" java class has been changed. The readable organizations were not taken into
account and that was the reason why the default value of the Unit of Measure was not taken into account when
a new product was created by the user with user level as "organization".

Now the readable organizatios are taken thanks to the "Utility.getContext(conn, vars, "#AccesibleOrgTree", "LoginHandler")"
code.
"#AccesibleOrgTree" has been put instead of "#User_Org", which is an string with the readable organizations inside.

Now when a user with user level as "organization" creates a new product, the default value of the UOM is
taken.

--HG--
extra : rebase_source : 7e3b882778de92d2bd4a5d79303ee9776e552366