Fixes bug 27992: Create Inventory Process should take into account attributes
'Create Inventory Count List' process in Physical Inventory it was not adding products with attribute set if the same product already exists in the physical inventory lines. This was because there was a select that was avoiding to select the products with different attribute set. The attribute_set_id has been added to this select to take into account these ones.