Fixes Bug 28441:Validate Costing Rule doesn't fail if there prev transactions.
The process was failing while trying to create Transaction Cost records in initializeOldTrx and updateInventoriesCostAndProcessInitInventories processes. It was not setting any value to accounting date column, which is mandatory. This column has been populated in the following way: If the related transaction is a shipment the accounting date of the shipment is assigned, if not, the movement date of the transaction will be assigned.

--HG--
extra : transplant_source : %17%AD%FF%E3%F2%CD%A3L%E6%8A%0CE%B8%FBZ%7F%C9Y%AF%E1