Related to Issue 28389: Fixed Posting of LC Cost with Invoice Line
While processing a Landed Cost with an Invoice in a Landed Cost Cost it, if the invoice has an exchange rate different from the system, a new Cost Adjustment is created with that difference between the rates. But this difference it was not taking into account while setting the matched amount in Landed Cost Cost, so the Post process it was not working well.

--HG--
extra : transplant_source : RE%8C%20%B3%C4%0E%07%1C9Xr%5D%20%84%9D%CC%9A%98%EB