Fixes Issue 26511: Display correct error message while deleting the orderline.
The problem on this issue is that the trigger is not able to delete the related reservation of the orderline and the given error is displayed. So in this case the error message displayed has been improved to warn the user that needs to remove the reservation manually before deleting the Sales Order Line.

--HG--
extra : rebase_source : c99471bf8dc11182025a99c343d5a4930119ed42